requirements.txt
discord.py
pyyaml
pytz
google-generativeai